requirements.txt
beautifulsoup4
comtypes
cssutils
einops
hdbscan
huggingface_hub
insightface
ipymarkup
Jinja2
natasha
numpy==1.26.4
onnxruntime-gpu
opencv-python
packaging
pandas
Pillow
plotly
photoshop-python-api
psd-tools
psutil
pydantic
pypiwin32
pymorphy3
PySide6
rawpy
regex
safetensors
scikit-learn
scipy
toml
torch
tqdm
transformers
transparent-background[gui]